requirements.txt
streamlit
python-dotenv
groq
sendgrid
pymupdf
reportlab
requests
beautifulsoup4
lxml
langchain
langchain-groq
langchain-google-genai
langgraph
resend
unidecode
python-jobspy>=1.1,<2
# ── Vector retrieval (e3) ──────────────────────────────────────
# Adds ~400 MB to the install footprint (PyTorch + MiniLM weights).
# Disable at runtime by setting USE_VECTOR_RETRIEVAL=0; the agent
# falls back to full-CV prompting with zero behaviour change.
chromadb>=0.4,<0.6
sentence-transformers>=2.2